+++ packages/backend/src/auth/auth.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
-import { GoogleAuthController } from '@/auth/googleAuth.controller';
-import { GoogleAuthService } from '@/auth/googleAuth.service';
-import { GoogleStrategy } from '@/auth/passport/google.strategy';
+import { GoogleAuthController } from '@/auth/google/googleAuth.controller';
+import { GoogleAuthService } from '@/auth/google/googleAuth.service';
+import { GoogleStrategy } from '@/auth/google/strategy/google.strategy';
 import { SessionSerializer } from '@/auth/passport/session.serializer';
 import { UserModule } from '@/user/user.module';
