+++ packages/backend/src/auth/google/googleAuth.controller.ts
@@ -1,8 +1,7 @@
 import { Controller, Get, Req, UseGuards } from '@nestjs/common';
 import { ApiOkResponse, ApiOperation, ApiTags } from '@nestjs/swagger';
 import { Request } from 'express';
-import { GoogleAuthGuard } from '@/auth/guard/google.guard';
-import { User } from '@/user/domain/user.entity';
+import { GoogleAuthGuard } from '@/auth/google/guard/google.guard';
 
 @ApiTags('Auth')
 @Controller('auth/google')
@@ -21,9 +20,10 @@ export class GoogleAuthController {
 
   @Get('/redirect')
   @UseGuards(GoogleAuthGuard)
-  async handleRedirect(@Req() request: Request) {
-    const user = request.user as User;
-    return { nickname: user.nickname, email: user.email };
+  async handleRedirect() {
+    return {
+      message: 'success google login',
+    };
   }
 
   @ApiOperation({
